+++ common/build.gradle
@@ -6,10 +6,10 @@ plugins {
     id 'io.spring.dependency-management'
 }
 bootJar {
-    enabled = true
+    enabled = false
 }
 jar {
-    enabled = false
+    enabled = true
 }
 dependencies {
     api 'org.springframework.boot:spring-boot-starter-web'
