+++ summarizer/src/main/resources/application.yml
@@ -1,18 +1,18 @@
 logging:
   level:
-    org.springframework.batch: DEBUG
+#    org.springframework.batch: DEBUG
     org.zerock.finedu: DEBUG
 spring:
   datasource:
     driver-class-name: com.mysql.cj.jdbc.Driver
-    url: jdbc:mysql://mysql:3306/finedu?useSSL=false&serverTimezone=UTC&characterEncoding=UTF-8
-    username: ${SPRING_DATASOURCE_USERNAME:finedu_user}
-    password: ${SPRING_DATASOURCE_PASSWORD:finedu_password}
-  batch:
-    job:
-      enabled: true  # 애플리케이션 시작 시 Batch Job 실행
-    jdbc:
-      initialize-schema: always  # Spring Batch 테이블 자동 생성
+    url: jdbc:mysql://mysql:3306/finedu?useSSL=false&allowPublicKeyRetrieval=true&serverTimezone=UTC&characterEncoding=UTF-8
+    username: ${MYSQL_USER:finedu_user}
+    password: ${MYSQL_PASSWORD:finedu_password}
+#  batch:
+#    job:
+#      enabled: true  # 애플리케이션 시작 시 Batch Job 실행
+#    jdbc:
+#      initialize-schema: always  # Spring Batch 테이블 자동 생성
   jpa:
     database-platform: org.hibernate.dialect.MySQLDialect
     show-sql: true
@@ -25,9 +25,9 @@ spring:
 
 openai:
   api-key: ${OPENAI_API_KEY}
+  model: gpt-3.5-turbo-0125
 
 server:
   port: ${SERVER_PORT:8081}
   address: 0.0.0.0
-
